Performing an error handling returning the status code 404, in case the user passes an invalid id.

+++ src/main/java/br/com/alura/forum/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.alura.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -58,25 +59,39 @@ public class TopicosController {
     }
 
     @GetMapping("/{id}")
-    public DetalhesTopicoDto detalhar(@PathVariable Long id) {
+    public ResponseEntity<DetalhesTopicoDto> detalhar(@PathVariable Long id) {
 
-        Topico topico = topicoRepository.getOne(id);
-        return new DetalhesTopicoDto(topico);
+        Optional<Topico> topico = topicoRepository.findById(id);
+
+        if (topico.isPresent()) { // existe um metodo no Optional is Present para verificar se esta presente
+            return ResponseEntity.ok(new DetalhesTopicoDto(topico.get())); // aqui é pra pegar o topico que esta dentro
+                                                                           // do Optional
+        }
+
+        return ResponseEntity.notFound().build(); // Retorna o erro 404
     }
 
     @PutMapping("/{id}")
     @Transactional // senao ele nao vai mandar o update no banco de dados
     public ResponseEntity<TopicoDto> atualizar(@PathVariable Long id, @RequestBody @Valid AtualizacaoTopicoForm form) {
-        Topico topico = form.atualizar(id, topicoRepository);
+        Optional<Topico> optional = topicoRepository.findById(id);
 
-        return ResponseEntity.ok(new TopicoDto(topico));
+        if (optional.isPresent()) {
+            Topico topico = form.atualizar(id, topicoRepository);
+            return ResponseEntity.ok(new TopicoDto(topico));
+        }
+        return ResponseEntity.notFound().build(); // Retorna o erro;
     }
 
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<?> remover(@PathVariable Long id) {
-        topicoRepository.deleteById(id);
+        Optional<Topico> optional = topicoRepository.findById(id);
 
-        return ResponseEntity.ok().build();
+        if (optional.isPresent()) {
+            topicoRepository.deleteById(id);
+            return ResponseEntity.ok().build();
+        }
+        return ResponseEntity.notFound().build(); // Retorna o erro;
     }
 }
